feat: SubscriptionRootResolverにroom_updatedメソッドを追加
