+++ src/resolver.rs
@@ -1,2 +1,3 @@
 pub mod mutation_root_resolver;
 pub mod query_root_resolver;
+pub mod subscription_root_resolver;
